+++ src/search/landmarks/landmark_factory_zhu_givan.cc
@@ -232,7 +232,7 @@ lm_set LandmarkFactoryZhuGivan::apply_operator_and_propagate_labels(
     lm_set precond_label_union = union_of_precondition_labels(op, current);
 
     const vector<Effect> &effects = op.get_effects();
-    for (int i = 0; i < effects.size(); ++i) {
+    for (size_t i = 0; i < effects.size(); ++i) {
         const int var = effects[i].var;
         const int post = effects[i].val;
