+++ src/search/state.cc
@@ -27,7 +27,7 @@ int State::operator[](int index) const {
 }
 
 void State::dump_pddl() const {
-    for (int i = 0; i < g_variable_domain.size(); ++i) {
+    for (size_t i = 0; i < g_variable_domain.size(); ++i) {
         const string &fact_name = g_fact_names[i][(*this)[i]];
         if (fact_name != "<none of those>")
             cout << fact_name << endl;
